Move FailWithLocC into its own module.
