+++ semantic-core/src/Analysis/Concrete.hs
@@ -13,8 +13,8 @@ import qualified Algebra.Graph as G
 import qualified Algebra.Graph.Export.Dot as G
 import           Analysis.Eval
 import           Control.Applicative (Alternative (..))
+import           Control.Carrier.Fail.WithLoc
 import           Control.Effect
-import           Control.Effect.Fail
 import           Control.Effect.Fresh
 import           Control.Effect.NonDet
 import           Control.Effect.Reader hiding (Local)
