+++ semantic-core/src/Analysis/ImportGraph.hs
@@ -8,8 +8,8 @@ module Analysis.ImportGraph
 import           Analysis.Eval
 import           Analysis.FlowInsensitive
 import           Control.Applicative (Alternative(..))
+import           Control.Carrier.Fail.WithLoc
 import           Control.Effect
-import           Control.Effect.Fail
 import           Control.Effect.Fresh
 import           Control.Effect.Reader
 import           Control.Effect.State
