+++ semantic-core/src/Analysis/Typecheck.hs
@@ -10,8 +10,8 @@ module Analysis.Typecheck
 import           Analysis.Eval
 import           Analysis.FlowInsensitive
 import           Control.Applicative (Alternative (..))
+import           Control.Carrier.Fail.WithLoc
 import           Control.Effect.Carrier
-import           Control.Effect.Fail
 import           Control.Effect.Fresh as Fresh
 import           Control.Effect.Reader hiding (Local)
 import           Control.Effect.State
