+++ semantic-core/src/Data/Loc.hs
@@ -1,21 +1,13 @@
-{-# LANGUAGE FlexibleInstances, GeneralizedNewtypeDeriving, MultiParamTypeClasses, OverloadedStrings, RecordWildCards, TypeOperators, UndecidableInstances #-}
+{-# LANGUAGE OverloadedStrings, RecordWildCards #-}
 module Data.Loc
 ( Loc(..)
 , interactive
 , here
 , stackLoc
-, FailWithLocC(..)
-, runFailWithLoc
 ) where
 
-import Control.Applicative
-import Control.Effect.Carrier
-import Control.Effect.Error
-import Control.Effect.Fail
-import Control.Effect.Reader
 import Data.Text (Text, pack)
 import GHC.Stack
-import Prelude hiding (fail)
 import Source.Span
 
 data Loc = Loc
@@ -38,19 +30,3 @@ stackLoc cs = case getCallStack cs of
 
 fromGHCSrcLoc :: SrcLoc -> Loc
 fromGHCSrcLoc SrcLoc{..} = Loc (pack srcLocFile) (Span (Pos srcLocStartLine srcLocStartCol) (Pos srcLocEndLine srcLocEndCol))
-
-
-runFailWithLoc :: FailWithLocC m a -> m (Either (Loc, String) a)
-runFailWithLoc = runError . runFailWithLocC
-
-newtype FailWithLocC m a = FailWithLocC { runFailWithLocC :: ErrorC (Loc, String) m a }
-  deriving (Alternative, Applicative, Functor, Monad)
-
-instance (Carrier sig m, Effect sig, Member (Reader Loc) sig) => MonadFail (FailWithLocC m) where
-  fail s = do
-    loc <- ask
-    FailWithLocC (throwError (loc :: Loc, s))
-
-instance (Carrier sig m, Effect sig, Member (Reader Loc) sig) => Carrier (Fail :+: sig) (FailWithLocC m) where
-  eff (L (Fail s)) = fail s
-  eff (R other)    = FailWithLocC (eff (R (handleCoercible other)))
